Moved internal/coreengine to service_packs/coreengine
