+++ service_packs/kubernetes/container_registry_access/container_registry_access.go
@@ -8,8 +8,8 @@ import (
 	"github.com/cucumber/godog"
 
 	"github.com/citihub/probr/internal/config"
-	"github.com/citihub/probr/internal/coreengine"
 	"github.com/citihub/probr/internal/utils"
+	"github.com/citihub/probr/service_packs/coreengine"
 	"github.com/citihub/probr/service_packs/kubernetes"
 )
