+++ service_packs/kubernetes/container_registry_access/helpers.go
@@ -4,8 +4,8 @@ import (
 	"strings"
 
 	"github.com/citihub/probr/internal/config"
-	"github.com/citihub/probr/internal/coreengine"
 	"github.com/citihub/probr/internal/summary"
+	"github.com/citihub/probr/service_packs/coreengine"
 	"github.com/citihub/probr/service_packs/kubernetes"
 	"github.com/cucumber/godog"
 	apiv1 "k8s.io/api/core/v1"
