+++ service_packs/kubernetes/general/general.go
@@ -12,8 +12,8 @@ import (
 	"github.com/cucumber/godog"
 
 	"github.com/citihub/probr/internal/config"
-	"github.com/citihub/probr/internal/coreengine"
 	"github.com/citihub/probr/internal/utils"
+	"github.com/citihub/probr/service_packs/coreengine"
 	"github.com/citihub/probr/service_packs/kubernetes"
 )
