+++ service_packs/kubernetes/general/helpers.go
@@ -1,8 +1,8 @@
 package general
 
 import (
-	"github.com/citihub/probr/internal/coreengine"
 	"github.com/citihub/probr/internal/summary"
+	"github.com/citihub/probr/service_packs/coreengine"
 	"github.com/citihub/probr/service_packs/kubernetes"
 	"github.com/cucumber/godog"
 )
