+++ service_packs/kubernetes/iam/helpers.go
@@ -7,8 +7,8 @@ import (
 	"log"
 
 	"github.com/citihub/probr/internal/config"
-	"github.com/citihub/probr/internal/coreengine"
 	"github.com/citihub/probr/internal/summary"
+	"github.com/citihub/probr/service_packs/coreengine"
 	"github.com/citihub/probr/service_packs/kubernetes"
 	"github.com/cucumber/godog"
