+++ service_packs/kubernetes/internet_access/helpers.go
@@ -6,8 +6,8 @@ import (
 	"strconv"
 
 	"github.com/citihub/probr/internal/config"
-	"github.com/citihub/probr/internal/coreengine"
 	"github.com/citihub/probr/internal/summary"
+	"github.com/citihub/probr/service_packs/coreengine"
 	"github.com/citihub/probr/service_packs/kubernetes"
 	"github.com/cucumber/godog"
 	apiv1 "k8s.io/api/core/v1"
