+++ service_packs/kubernetes/kubernetes.go
@@ -9,9 +9,9 @@ import (
 	apiv1 "k8s.io/api/core/v1"
 
 	"github.com/citihub/probr/internal/config"
-	"github.com/citihub/probr/internal/coreengine"
 	"github.com/citihub/probr/internal/summary"
 	"github.com/citihub/probr/internal/utils"
+	"github.com/citihub/probr/service_packs/coreengine"
 )
 
 var AssetsDir string
