+++ service_packs/kubernetes/pack/kubernetes_pack_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/citihub/probr/internal/config"
-	"github.com/citihub/probr/internal/coreengine"
+	"github.com/citihub/probr/service_packs/coreengine"
 )
 
 func TestGetProbes(t *testing.T) {
