+++ service_packs/storage/azure/encryption_at_rest/encryption_at_rest.go
@@ -6,8 +6,8 @@ import (
 	"log"
 	"strings"
 
-	"github.com/citihub/probr/internal/coreengine"
 	"github.com/citihub/probr/internal/summary"
+	"github.com/citihub/probr/service_packs/coreengine"
 	"github.com/cucumber/godog"
 )
