+++ service_packs/storage/pack/pack.go
@@ -2,7 +2,7 @@ package storage_pack
 
 import (
 	"github.com/citihub/probr/internal/config"
-	"github.com/citihub/probr/internal/coreengine"
+	"github.com/citihub/probr/service_packs/coreengine"
 	azure_access_whitelisting "github.com/citihub/probr/service_packs/storage/azure/access_whitelisting"
 	azure_encryption_at_rest "github.com/citihub/probr/service_packs/storage/azure/encryption_at_rest"
 	azure_encryption_in_flight "github.com/citihub/probr/service_packs/storage/azure/encryption_in_flight"
